renamed alternativenoresults to noresults
